:art: Refactored a bit
