+++ packages/cli/src/utils/i18n.ts
@@ -1,0 +1,14 @@
+import * as fs from "fs";
+import path from "path";
+
+export async function* getFiles(dir: string): any {
+  const dirents = await fs.readdirSync(dir, { withFileTypes: true });
+  for (const dirent of dirents) {
+    const res = path.resolve(dir, dirent.name);
+    if (dirent.isDirectory()) {
+      yield* getFiles(res);
+    } else {
+      yield res;
+    }
+  }
+}
